requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
frozenlist==1.3.3
html2text==2020.1.16
idna==3.4
importlib-metadata==6.6.0
lxml==4.9.2
multidict==6.0.4
openai==0.27.5
pdfminer.six==20221105
pycparser==2.21
python-docx==0.8.11
requests==2.29.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.9.2
zipp==3.15.0
